fix: Fix latency bug
With a B210, it was likely to run into issues with latency. We saw an
issue with high latency at around ~24ms. However, less than 3ms is
expected. This behavior can be observed with a N310.

`set_time_now()` instead of unknown PPS seems to have fixed it. A loop
to resync during longer runs seems to be highly appropriate.

Signed-off-by: Johannes Demel <demel@ant.uni-bremen.de>